backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Retry Logic
tenacity==8.2.3

# Environment
python-dotenv==1.0.1

# Caching (optional)
redis==5.0.1

# Google Ads Transparency Scraper
Google-Ads-Transparency-Scraper==1.6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Linting (dev)
ruff==0.2.1
mypy==1.8.0
